Split bind_copy and bind_read into multiple files, remove CALL read_pandas
