+++ src/binder/bind/CMakeLists.txt
@@ -1,12 +1,12 @@
 add_subdirectory(copy)
 add_subdirectory(ddl)
+add_subdirectory(read)
 
 add_library(
         kuzu_binder_bind
         OBJECT
         bind_attach_database.cpp
         bind_comment_on.cpp
-        bind_copy.cpp
         bind_create_macro.cpp
         bind_ddl.cpp
         bind_detach_database.cpp
