+++ src/binder/bind/read/bind_match.cpp
@@ -1,0 +1,67 @@
+#include "binder/binder.h"
+#include "binder/query/reading_clause/bound_match_clause.h"
+#include "parser/query/reading_clause/match_clause.h"
+
+using namespace kuzu::common;
+using namespace kuzu::parser;
+
+namespace kuzu {
+namespace binder {
+
+std::unique_ptr<BoundReadingClause> Binder::bindMatchClause(const ReadingClause& readingClause) {
+    auto& matchClause = readingClause.constCast<MatchClause>();
+    auto boundGraphPattern = bindGraphPattern(matchClause.getPatternElementsRef());
+    if (matchClause.hasWherePredicate()) {
+        boundGraphPattern.where = bindWhereExpression(*matchClause.getWherePredicate());
+    }
+    rewriteMatchPattern(boundGraphPattern);
+    auto boundMatch = std::make_unique<BoundMatchClause>(
+        std::move(boundGraphPattern.queryGraphCollection), matchClause.getMatchClauseType());
+    boundMatch->setPredicate(boundGraphPattern.where);
+    return boundMatch;
+}
+
+void Binder::rewriteMatchPattern(BoundGraphPattern& boundGraphPattern) {
+    // Rewrite self loop edge
+    // e.g. rewrite (a)-[e]->(a) as [a]-[e]->(b) WHERE id(a) = id(b)
+    expression_vector selfLoopEdgePredicates;
+    auto& graphCollection = boundGraphPattern.queryGraphCollection;
+    for (auto i = 0u; i < graphCollection.getNumQueryGraphs(); ++i) {
+        auto queryGraph = graphCollection.getQueryGraphUnsafe(i);
+        for (auto& queryRel : queryGraph->getQueryRels()) {
+            if (!queryRel->isSelfLoop()) {
+                continue;
+            }
+            auto src = queryRel->getSrcNode();
+            auto dst = queryRel->getDstNode();
+            auto newDst = createQueryNode(dst->getVariableName(), dst->getTableIDs());
+            queryGraph->addQueryNode(newDst);
+            queryRel->setDstNode(newDst);
+            auto predicate = expressionBinder.createEqualityComparisonExpression(
+                src->getInternalID(), newDst->getInternalID());
+            selfLoopEdgePredicates.push_back(std::move(predicate));
+        }
+    }
+    auto where = boundGraphPattern.where;
+    for (auto& predicate : selfLoopEdgePredicates) {
+        where = expressionBinder.combineBooleanExpressions(ExpressionType::AND, predicate, where);
+    }
+    // Rewrite key value pairs in MATCH clause as predicate
+    for (auto i = 0u; i < graphCollection.getNumQueryGraphs(); ++i) {
+        auto queryGraph = graphCollection.getQueryGraphUnsafe(i);
+        for (auto& pattern : queryGraph->getAllPatterns()) {
+            for (auto& [propertyName, rhs] : pattern->getPropertyDataExprRef()) {
+                auto propertyExpr =
+                    expressionBinder.bindNodeOrRelPropertyExpression(*pattern, propertyName);
+                auto predicate =
+                    expressionBinder.createEqualityComparisonExpression(propertyExpr, rhs);
+                where = expressionBinder.combineBooleanExpressions(ExpressionType::AND, predicate,
+                    where);
+            }
+        }
+    }
+    boundGraphPattern.where = std::move(where);
+}
+
+} // namespace binder
+} // namespace kuzu
