+++ test/test_files/exceptions/binder/binder_error.test
@@ -229,7 +229,7 @@ Binder exception: The number of columns to union/union all must be the same.
 -LOG UnionAllUnmatchedDataTypesOfExpressions
 -STATEMENT MATCH (p:person) RETURN p.fName UNION ALL MATCH (p1:person) RETURN p1.age
 ---- error
-Binder exception: p1.age has data type INT64 but (STRING) was expected.
+Binder exception: p1.age has data type INT64 but STRING was expected.
 
 -LOG UnionAndUnionAllInSingleQuery
 -STATEMENT MATCH (p:person) RETURN p.age UNION ALL MATCH (p1:person) RETURN p1.age UNION MATCH (p1:person) RETURN p1.age
@@ -444,7 +444,7 @@ Catalog exception: Table: person1 does not exist.
 -LOG DropNonExistedColumn
 -STATEMENT alter table person drop random
 ---- error
-Binder exception: person table doesn't have property random.
+Binder exception: person table does not have property random.
 
 -LOG DropPrimaryKeyColumn
 -STATEMENT alter table person drop ID
@@ -535,7 +535,7 @@ Binder exception: Invalid option name: thread.
 -LOG InvalidCallOptionValue
 -STATEMENT CALL threads='abc'
 ---- error
-Binder exception: abc has data type STRING but (INT64) was expected.
+Binder exception: abc has data type STRING but INT64 was expected.
 
 -LOG AllShortestPathInvalidLowerBound
 -STATEMENT MATCH p = (a)-[* ALL SHORTEST 2..3]-(b) RETURN p
