+++ test/test_files/tinysnb/function/start_end_node.test
@@ -56,7 +56,7 @@
 -LOG StartNodeTestRecursiveRel
 -STATEMENT MATCH (:person { fName: "Alice" })-[r:knows *1..2]->(friend:person) RETURN START_NODE(r)
 ---- error
-Binder exception: r has data type RECURSIVE_REL but (REL) was expected.
+Binder exception: r has data type RECURSIVE_REL but REL was expected.
 
 
 -CASE FunctionEndNode
@@ -112,4 +112,4 @@ Binder exception: r has data type RECURSIVE_REL but (REL) was expected.
 -LOG EndNodeTestRecursiveRel
 -STATEMENT MATCH (:person { fName: "Alice" })-[r:knows *1..2]->(friend:person) RETURN END_NODE(r)
 ---- error
-Binder exception: r has data type RECURSIVE_REL but (REL) was expected.
+Binder exception: r has data type RECURSIVE_REL but REL was expected.
