+++ test/test_files/transaction/create_macro/create_macro.test
@@ -152,4 +152,4 @@ result|6|5.600000|00:20:00|[1,3,2]
 ---- ok
 -STATEMENT CREATE MACRO var_macro(x) AS x
 ---- error
-Can't execute a write query inside a read-only transaction.
+Can not execute a write query inside a read-only transaction.
