+++ tools/python_api/test/test_df_pyarrow.py
@@ -71,7 +71,7 @@ def pyarrow_test_helper(establish_connection, n, k):
     df = generate_primitive_df(n, names, schema).sort_values(by=['int32col', 'int64col', 'uint64col', 'floatcol'])
     patable = pa.Table.from_pandas(df).select(names)
     result = conn.execute(
-        'CALL READ_PANDAS(df) RETURN boolcol, int32col, int64col, uint64col, floatcol ORDER BY int32col, int64col, uint64col, floatcol'
+        'LOAD FROM df RETURN boolcol, int32col, int64col, uint64col, floatcol ORDER BY int32col, int64col, uint64col, floatcol'
     ).get_as_arrow(n)
     if (not tables_equal(patable, result)):
         print(patable)
@@ -127,7 +127,7 @@ def test_pyarrow_time(conn_db_readonly : ConnDB) -> None:
         'col10': arrowtopd(col10)
         #'col11': arrowtopd(col11)
     })
-    result = conn.execute('CALL READ_PANDAS(df) RETURN *').get_as_df()
+    result = conn.execute('LOAD FROM df RETURN *').get_as_df()
     for colname in ['col1', 'col2', 'col3']:
         for expected, actual in zip(df[colname], result[colname]):
             tmp1 = expected if type(expected) is timedelta else expected.to_pytimedelta()
@@ -166,7 +166,7 @@ def test_pyarrow_blob(conn_db_readonly : ConnDB) -> None:
         'col3' : arrowtopd(col3),
         'col4' : arrowtopd(col4)
     }).sort_values(by=['index'])
-    result = conn.execute('CALL READ_PANDAS(df) RETURN * ORDER BY index').get_as_df()
+    result = conn.execute('LOAD FROM df RETURN * ORDER BY index').get_as_df()
     for colname in ['col1', 'col2', 'col3', 'col4']:
         for expected, actual in zip(df[colname], result[colname]):
             if is_null(expected) or is_null(actual):
@@ -203,7 +203,7 @@ def test_pyarrow_string(conn_db_readonly : ConnDB) -> None:
         'col2' : arrowtopd(col2),
         'col3' : arrowtopd(col3),
     }).sort_values(by=['index'])
-    result = conn.execute('CALL READ_PANDAS(df) RETURN * ORDER BY index').get_as_df()
+    result = conn.execute('LOAD FROM df RETURN * ORDER BY index').get_as_df()
     for colname in ['col1', 'col2', 'col3']:
         for expected, actual in zip(df[colname], result[colname]):
             if is_null(expected) or is_null(actual):
@@ -228,7 +228,7 @@ def test_pyarrow_dict(conn_db_readonly : ConnDB) -> None:
         'col1' : arrowtopd(col1),
         'col2' : arrowtopd(col2)
     })
-    result = conn.execute('CALL READ_PANDAS(df) RETURN * ORDER BY index').get_as_df()
+    result = conn.execute('LOAD FROM df RETURN * ORDER BY index').get_as_df()
     for colname in ['col1', 'col2']:
         for expected, actual in zip(df[colname], result[colname]):
             assert expected == actual
@@ -248,7 +248,7 @@ def test_pyarrow_list(conn_db_readonly : ConnDB) -> None:
         'col1': arrowtopd(col1),
         'col2': arrowtopd(col2)
     })
-    result = conn.execute('CALL READ_PANDAS(df) RETURN * ORDER BY index')
+    result = conn.execute('LOAD FROM df RETURN * ORDER BY index')
     idx = 0
     while result.has_next():
         assert idx < len(index)
@@ -276,7 +276,7 @@ def test_pyarrow_struct(conn_db_readonly : ConnDB) -> None:
         'index': arrowtopd(index),
         'col1': arrowtopd(col1)
     })
-    result = conn.execute('CALL READ_PANDAS(df) RETURN * ORDER BY index')
+    result = conn.execute('LOAD FROM df RETURN * ORDER BY index')
     idx = 0
     while result.has_next():
         assert idx < len(index)
@@ -302,7 +302,7 @@ def test_pyarrow_union(conn_db_readonly : ConnDB) -> None:
         'index': arrowtopd(index),
         'col1': arrowtopd(col1)
     })
-    result = conn.execute('CALL READ_PANDAS(df) RETURN * ORDER BY index')
+    result = conn.execute('LOAD FROM df RETURN * ORDER BY index')
     idx = 0
     while result.has_next():
         assert idx < len(index)
@@ -328,7 +328,7 @@ def test_pyarrow_map(conn_db_readonly : ConnDB) -> None:
     df = pd.DataFrame({
         'index': arrowtopd(index),
         'col1': arrowtopd(col1)})
-    result = conn.execute('CALL READ_PANDAS(df) RETURN * ORDER BY index')
+    result = conn.execute('LOAD FROM df RETURN * ORDER BY index')
     idx = 0
     while result.has_next():
         assert idx < len(index)
