+++ tools/python_api/test/test_scan_pandas.py
@@ -192,10 +192,8 @@ def test_scan_pandas(tmp_path: Path) -> None:
     }
     df = pd.DataFrame(data)
     df["datetime_microseconds_tz"] = df["datetime_microseconds_tz"].dt.tz_localize("US/Eastern")
-    results = conn.execute("CALL READ_PANDAS(df) RETURN *")
+    results = conn.execute("LOAD FROM df RETURN *")
     validate_scan_pandas_results(results)
-    results2 = conn.execute("LOAD FROM df RETURN *")
-    validate_scan_pandas_results(results2)
 
 
 def test_scan_pandas_timestamp(tmp_path: Path) -> None:
@@ -213,7 +211,7 @@ def test_scan_pandas_timestamp(tmp_path: Path) -> None:
     df = pd.DataFrame({"timestamp": ts})
     # Pandas automatically converts the column from object to timestamp, so we need to manually cast back to object.
     df = df.astype({"timestamp": "object"}, copy=False)
-    results = conn.execute("CALL READ_PANDAS(df) RETURN *")
+    results = conn.execute("LOAD FROM df RETURN *")
     assert results.get_next() == [datetime.datetime(1996, 2, 15, hour=12, minute=22, second=54)]
     assert results.get_next() == [datetime.datetime(2011, 3, 11, minute=11, hour=5)]
     assert results.get_next() == [None]
@@ -262,7 +260,7 @@ def test_scan_pandas_with_filter(tmp_path: Path) -> None:
     df = pd.DataFrame(data)
     # Dummy query to ensure the READ_PANDAS function is persistent after a write transaction.
     conn.execute("CREATE NODE TABLE PERSON1(ID INT64, PRIMARY KEY(ID))")
-    results = conn.execute("CALL READ_PANDAS(df) WHERE id > 20 RETURN id + 5, weight, name")
+    results = conn.execute("LOAD FROM df WHERE id > 20 RETURN id + 5, weight, name")
     assert results.get_next() == [27, 23.2, "ñ"]
     assert results.get_next() == [105, 42.9, "😊"]
 
@@ -277,7 +275,7 @@ def test_large_pd(tmp_path: Path) -> None:
         "odd": np.array(odd_numbers, dtype=np.int64),
         "even": np.array(even_numbers, dtype=np.int64),
     })
-    result = conn.execute("CALL READ_PANDAS(df) RETURN *").get_as_df()
+    result = conn.execute("LOAD FROM df RETURN *").get_as_df()
     assert result["odd"].to_list() == odd_numbers
     assert result["even"].to_list() == even_numbers
 
@@ -299,7 +297,7 @@ def test_pandas_scan_demo(tmp_path: Path) -> None:
     person = pd.DataFrame({"id": id, "age": age, "height": height_in_cm, "is_student": is_student})
 
     result = conn.execute(
-        "CALL READ_PANDAS(person) with avg(height / 2.54) as height_in_inch MATCH (s:student) WHERE s.height > "
+        "LOAD FROM person with avg(height / 2.54) as height_in_inch MATCH (s:student) WHERE s.height > "
         "height_in_inch RETURN s"
     ).get_as_df()
     assert len(result) == 2
@@ -308,7 +306,7 @@ def test_pandas_scan_demo(tmp_path: Path) -> None:
 
     conn.execute("CREATE NODE TABLE person(ID INT64, age UINT16, height UINT32, is_student BOOLean, PRIMARY KEY(ID))")
     conn.execute(
-        "CALL READ_PANDAS(person) CREATE (p:person {ID: id, age: age, height: height, is_student: is_student})"
+        "LOAD FROM person CREATE (p:person {ID: id, age: age, height: height, is_student: is_student})"
     )
     result = conn.execute("MATCH (p:person) return p.*").get_as_df()
     assert np.all(result["p.ID"].to_list() == id)
@@ -334,7 +332,7 @@ def test_scan_all_null(tmp_path: Path) -> None:
     conn = kuzu.Connection(db)
     data = {"id": np.array([None, None, None], dtype=object)}
     df = pd.DataFrame(data)
-    result = conn.execute("CALL READ_PANDAS(df) RETURN *")
+    result = conn.execute("LOAD FROM df RETURN *")
     assert result.get_next() == [None]
     assert result.get_next() == [None]
     assert result.get_next() == [None]
